Updated Tinybird service to use wrapper pattern
